added death and fixed player catching platforms in level1
added the Die function. Boots are now used to detect whether the player is jumping on an enemy or just touching an enemy. If the latter is true, the player will lose a hit point.
Added a pain tag for non-enemy objects that can inflict pain.
Still want to add recoil to enemy jumping. Should also look into invincibility period. Will likely need to update health monitoring to do so.